website-backend/requirements.txt
Flask==1.1.2
gdown==4.2.0
numpy==1.20.3
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
Werkzeug==2.0.2
gunicorn